+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add tooltip to disabled visibility selector.
 - Set visibility to "Followers" when self-replying to followers-only post.
 - Enable username autocomplete on login form.
+- Don't allow to edit posts with text/html content type.
 
 ### Fixed